Issue #1

+++ .scripts/checkstyle-pre-commit.sh
@@ -1,11 +1,9 @@
 #!/bin/bash
 
-# TODO
-#  Check if checkstyle jar exists in .checkstyle already. download if not. same with google_checks.
-#  then lint src/ or all of all
-
+# Checkstyle jar artifact
 CHECKSTYLE_VERSION="https://github.com/checkstyle/checkstyle/releases/download/checkstyle-10.12.0/checkstyle-10.12.0-all.jar"
 
+# File locations
 PRE_COMMIT_DIR=~/.cache/pre-commit/checkstyle
 CHECKSTYLE_JAR="checkstyle.jar"
 GOOGLE_CHECKS="google_checks.xml"
@@ -13,6 +11,7 @@ SUN_CHECKS="sun_checks.xml"
 OUTPUT_CACHE="${PRE_COMMIT_DIR}/output_cache.txt"
 OUTPUT_FILE="${PRE_COMMIT_DIR}/output.txt"
 
+# script flags and codes
 STRICT=false
 EXIT_CODE=0
 
@@ -25,31 +24,35 @@ do
     esac
 done
 
+# if error on script call.
 if [ $EXIT_CODE != 0 ]
 then
   echo "exit"
   exit $EXIT_CODE
 fi
 
-
+# make checkstyle dir in pre commit if does not exist
 if [ ! -d "${PRE_COMMIT_DIR}" ]
 then
   echo "caching directory under: ${PRE_COMMIT_DIR}"
   mkdir "${PRE_COMMIT_DIR}"
 fi
 
+# download checkstyle jar if not found
 if [ ! -f "${PRE_COMMIT_DIR}/${CHECKSTYLE_JAR}" ]
 then 
   echo "downloading checkstyle"
   curl -o "${PRE_COMMIT_DIR}/${CHECKSTYLE_JAR}" -LJO "${CHECKSTYLE_VERSION}"
 fi
 
+# download google checks config if not found
 if [ ! -f "${PRE_COMMIT_DIR}/${GOOGLE_CHECKS}" ]
 then
   echo "downloading google_checks config"
   curl -o "${PRE_COMMIT_DIR}/${GOOGLE_CHECKS}" https://raw.githubusercontent.com/checkstyle/checkstyle/master/src/main/resources/google_checks.xml
 fi
 
+# download sun checks config if not found
 if [ ! -f "${PRE_COMMIT_DIR}/${SUN_CHECKS}" ]
 then
   echo "downloading sun_checks config"
@@ -66,18 +69,23 @@ case "${CONFIG_ARG}" in
 
 esac
 
+# determine config path
 CHECKSTYLE_CONFIG="${PRE_COMMIT_DIR}/${CONFIG_ARG}"
 echo "running checkstyle using ${CONFIG_ARG} config"
 
 # HERE
 # clear old output file.
-rm "${OUTPUT_FILE}"
+if [ -f "${OUTPUT_FILE}" ]
+then
+  rm "${OUTPUT_FILE}"
+fi
 
 # iterate through checked in java files.
-ALL_FILES=$(git diff --cached --name-only)
+ALL_FILES=$(git diff --cached --name-status)
+FILE_STATUS=false
 for FILE in $ALL_FILES
-do 
-  if [[ "${FILE}" == *.java ]]
+do
+  if [[ "${FILE}" == *.java && $FILE_STATUS = true ]]
   then
     # HERE
     # run checkstyle command on specific file.
@@ -86,12 +94,24 @@ do
 
     cat "${OUTPUT_CACHE}" >> "${OUTPUT_FILE}"
   fi
+
+  # set status to true if positive change on file.
+  if [[ "${FILE}" = "D" || "${FILE}" = 'R'* ]]
+    then
+      FILE_STATUS=false
+    else
+      FILE_STATUS=true
+  fi
 done
 
+# HERE
 # cleanup output cache
-rm "${OUTPUT_CACHE}"
+if [ -f "${OUTPUT_CACHE}" ]
+then
+  rm "${OUTPUT_CACHE}"
+fi
 
-# sift through total output.
+# Filter all checkstyle output
 ERRORS=0
 while read line; do
   # parse output lines
@@ -101,6 +121,7 @@ while read line; do
     echo "${line}"
   fi
 
+  # if STRICT set WARN as error
   if [[ "${line}" == '[WARN]'* && $STRICT = true ]]
   then
     ERRORS=$((ERRORS + 1))
